Add files via upload
Consumer complaint resolution is important to any business. In this particular case we have been
given details consumer complaints along with whether consumer disputed with the conclusion. If
we are able to predict this , consumer likely disputed can be given more attention as to how the
complaints are handled as well as how to convincingly convey the final conclusions to them.